+++ src/resources/views/admin/employees/create.blade.php
@@ -118,10 +118,8 @@
                     @endif
 
                     <div class="mb-3 mt-3">
-                        @isset($departments)
-                            <label>{{ config("site-staff.siteDepartmentName") }}:</label>
-                           @include("site-staff::admin.departments.includes.tree-checkbox", ['departments' => $departments])
-                        @endisset
+                        <label>{{ config("site-staff.siteDepartmentName") }}:</label>
+                        @include("site-staff::admin.departments.includes.tree-checkbox")
                     </div>
 
                     <div class="btn-group mt-2"