+++ src/routes/site/employee.php
@@ -1,0 +1,12 @@
+<?php
+
+use Illuminate\Support\Facades\Route;
+
+Route::group([
+"namespace" => "App\Http\Controllers\Vendor\SiteStaff\Site",
+"middleware" => ["web"],
+"as" => "site.employees.",
+"prefix" => config("site-staff.employeeUrlName"),
+], function () {
+Route::get("/{employee}", "StaffEmployeeController@show")->name("show");
+});